covert unsigned int type to size_t in class definition fullIndex data member was size_t but in implementation (Movie.cpp file) it was unsiged int type. unsigned int Movie::fullIndex = -1; (it has no any error in Windows but it was compiler error in linux and wsl)

+++ src/movie.cpp
@@ -6,7 +6,7 @@
 #include <time.h>
 using namespace std;
 
-unsigned int Movie::fullIndex = -1;     //There is no movie 
+size_t Movie::fullIndex = -1;           //There is no movie 
                                         //(Because index 0 itself means the index of a house)
